Make AWSD + mouse the default conctrol scheme
Weirdly, the game is using the menu text for direct storage of the assigned keys,
such that the assignments live "twice", once in the menu text and once in controlloop.
Left mouse button is fire now, right mouse button changes weapons.

When you first start the Controls menu, it'll ignore the master assignments and
use whatever is set in  'KEY_LINES' and 'KEY_LINES2'. Upon exiting the menu,
it'll overwrite the master settings from whatever is stored in the menu.

Fix this by loading KEY_LINES and KEY_LINES2 with the contents of the master assignment
table.