Use same rotation math when map is drawn or not
During map drawing all level points will be transformed instead of only a subset.
This code seemed to try using some shortcuts(?) at the cost of slighly lower precision(?).
This caused the rendering to "warp" slighly when switching the map on or off.

Fixed by using the same math for both. Maybe revisit later and try to further optimize.