Open invisible window for menu screen.
This allows me to clear the mouse coursor.
Later on this may help with input handling.

This excacerbated the issue with the cleared backround as
now the window clears the background first.
The way to fix this would be to re-initialize the
Bitmap after opening screen and window somehow.
But that would require a temporary "scratch" Bitmap
for the screen to use before switching to our own...