Make window small
Opening the window clears its contents. We don't want that, so open it small.

The idea was to then resize it without responding to refresh requests
(since we're using simple refresh). But that didn't actually work.
I think Intuition did delay the clear post our setup and thus still managed
to clear the screen behind the window.