Another attempt to not make Intuition clear the title bar.
When opening the screen, provide a Bitmap setup in such a way that
a clear would only touch scratch memory (the planes used for the fire effect).
Then, after opening the screen, adjust the Bitmap 'for real'.
For some reason, that doesn't work. I think Intuition doesn't immediately
clear the title bar, but some time later.