Use Intuition screen for main rendering
Open a regular Intution screen with custom Bitmap for rendering.
This removes a ton of copper list and AB3D1 "old menu" and old HUD related stuff,
as well as a copperlist for the level-blurb text screen and endgame scroll.

The main copperlist "bigfield" would set the screen mode, setup a palette,
clear sprites, setup the modulo for individual lines 0-232 (for "double height" rendering)
as well as attach a separate bitmap to the very bottom, along with scrolling
for the status text rendering.
Various hardware register accesses (in particular DMACON/INTENA) were removed as well.

For now, we just restored the regular rendering screen.
The doubleheight rendering and scrolling status text will have to be replaced as
well, potentially with an Intuition-friendly user copperlist.

Various fade-ins and fade-outs have been disabled, as well as
the player-got-hit "redshift". Needs to be fixed in future commits.

Audio, input and serial related direct HW access is still left intact. Maybe
I can just get away with acquiring the corresponding resource.

An interesting findingis that large parts of the main rendering loop seem to run
in the VBLank (formerly copper interrupt), but moving it out of there into
the main game loop made everything run in slow motion.

Also TODO: menuscreen and mainscreen don't have to be separate and actually could share
their bitmaps. Need to store the menu background in fastmem and restore it when
menu opens. This also solves the issue with the Screen titlebar clearing the screen.