Use OS friendly user-copperlist to achieve doubleheight rendering
The copper list scans out even lines twice and skips odd lines.
The modulos are hard coded to expect FMODE 3.
The use of VideoControl was support to help later with
potentially doign screen dragging on the screen, but currently
it leads to crashes.

Move most screen related  code to its own file.