Use OneOverN table multiply in more places
...in particular in gouraud floor drawing!

This code is used to project the current line on screen onto the floor/ceiling.
The result is the distance of the intersection to the viewer, thus it is used to derived the
line of the floor visible through the line on screen.

The incoming value is already shifted by 6 bits, the values in the table are shifted by 14 bits;
moving the result out by 20 bits. This is not ideal. OneOverN can't use much less bits as
it would loose precision near the player. Maybe the incoming floor distance doesn't need
shifting by 6bits anymore?