Fix DOUBLEHEIGHT rendering for mouselook
mouselook alllows to shift the middle of the projection (SMIDDLEY) up and down,
which also moves the 'BIGMIDDLEY' offset up and down in the renderbuffer, which determines
where floor lines get drawn on screen.

The original DOUBLEHEIGHT floor rendering code assumed that the middle of the screen
was even-aligned for floors and odd for ceilings.
But moving BIGMIDDLEY up and down makes the assumptions invalid and we could end
up rendering to odd lines only, which in turn are skipped during C2P and scanout -
making all floors and ceilings vanish.

Fix by always aligning the verticel screen center to an even line in DOUBLEHEIGHT mode.
This might not be a 100% correct fix as typically moving the destination address would
require to also compensate sampling the floor texture. Thus we likely introduce a
slight wobble.

Also fix map rendering aspect ratio for 2xh rendering. The map lines are not correctly
rendered, since the lines are drawn at full resolution.

Ultimately, the double height rendering mode (via skipped lines) should die and be properly
replaced by just rendering a smaller image.