Fix issue with floor rendering
The optimization to use the OneOverN table instead of a slow divs.l division
created an issue in rooms with a wide open area where floors and ceilings
started to glitch out.
This was due to values after the multiplication with OneOverN exceeding 32bits,
leaving us with only the lower bits in the result.

Solve this by pre-shifting one of the factors down before the multiplication.