Hack menu to work on RTG
The original menu code will only redraw when there's some action
like cursor movement happening. The continuous update of the fire effect
and background scroll was done in the VBlank interrupt.
Thus, there's no particular "render menu frame" function available to hook into.

In order to hack RTG into the menu code, we let it still compose into
a regular 8 bit chipmem based bitmap, including the blitter-based fire effect.
To show up in the RTG screen, we ned to blit the bitmap into the RTG
screen (performing a planar-to-chunky conversion on the way) on a regular basis.

I couldn't find a suitable place to hook the RTG blit into for reasons
explained above. Kicking off the blit from the fake 'VBlank' timer interrupt
doesn't work and would likely not be legal.
So what I ended up  doing is to create a separate task that will get signaled
from the vblank interrupt and will perform the blit upon being signaled.

This is just a hack to get the menu working and be able to proceed to
make the RTG changes to the actual game. The whole menu system is a
hot mess and should be redone properly later on.

Using this blitting from chipmem into an RTG is extremely slow and
seems to eat up lots of performance, making the inital loading rather slow.